Only show download zips based on an allow list
- Move away from blindly showing all zips found on DHE. Use a hardcoded
list of zips that we expect to be listed on the website.

- It is tricky to construct the allowed list for the Betas zips because
beta zips are labeled by the liberty version. This is handled by a
function to construct the allow list on the fly.

- There can be more than one zip per beta version. Replace the previous
fix from c0f13d4b55a62fb55833aaa0a884966ff04ae78a with the new allow
list logic.