Update to sort based on version.
In 22.0.0.2 the uploads were changed from being in a date based directory to a version
based one. This messed up the sort logic which assumed that the directory was the date of the
release and it means that 22.0.0.10 no longer sorted as the most recent release.

Instead updating so if the version field is set we will sort based on version. The ideal is
we will covert each version section to be a three digit version with leading zeros. That
means that as long as we don't have 1000 numbers in any component then the sort will be fine.
Based on current versioning this will last until the year 3000 or until we do 3 releases a day
every day for a year, either seem to be unlikely.